Refactor routing and add new timing functionalities by consolidating routers, implementing new timing logic, and enhancing token verification in the API
